change i64 and u64 to long long for Windows compat

+++ dznl/NumericTypes.hpp
@@ -7,11 +7,11 @@ namespace dznl {
 using i8 = signed char;
 using i16 = signed short int;
 using i32 = signed int;
-using i64 = signed long int;
+using i64 = signed long long int;
 using u8 = unsigned char;
 using u16 = unsigned short int;
 using u32 = unsigned int;
-using u64 = unsigned long int;
+using u64 = unsigned long long int;
 using f32 = float;
 using f64 = double;
